fix: route to services, added team, fixed services
