+++ ROADMAP.md
@@ -1,0 +1,10 @@
+- [x] Change services
+- [x] Add team page
+- [x] Fill team page
+- [ ] Change `/projects/` route to `/services/`
+
+- [ ] Send messages to our email (contact form)
+
+# Future
+
+- [ ] Change blog posts
