--- contents/projects/hacksheffield.md
@@ -1,31 +1,0 @@
----
-title: HackSheffield
-titleShort: ''
-date: 2019-04-01
-image: '/assets/images/dynamic/hacksheffield/committee.jpg'
-description: Hackathon society which I have been involved in since starting university. See which hackathons I have attended and organised
-tags:
-  - Hackathons
-  - Event management
-  - Leadership
-  - Public speaking
-  - Travels
----
-
-The first hackathon I attended was HackSheffield 2.0, less than a month after I started university --- I had a fantastic time at the event, making <nuxt-link to="/projects/hype/">an Android app called Hype</nuxt-link>, so I decided to become more involved with hackathons and in particular, HackSheffield society.
-
-## Organising HackSheffield
-
-[HackSheffield](https://hacksheffield.co) is a society at the University of Sheffield dedicated to organising hackathons and sending members of the society across Europe to attend other hackathons. After HackSheffield 2.0, I joined the organising committee for HackSheffield, which at the time consisted of five Computer Science students. The first hackathon I organised was Local Hack Day 2016, coordinated by [Major League Hacking](https://mlh.io).
-
-Less than a year after attending my first hackathon, I was elected the President of HackSheffield, ensuring that everything was ran smoothly by the team and directing the vision of the society. As President of the society, one of my priorities was to get more people involved in hackathons, so I championed a series of workshops at the University of Sheffield --- entitled "An Introduction to Hackathons" --- to teach people some technology which they might find useful in a hackathon environment. I stepped down from the role as President in 2018 and ran instead for the role of Technical Officer. As Technical Officer, I was in charge of the society's infrastructure, including the society's website, email server and the society's many accounts on various platforms. I feel massively privileged to have played such a big part in shaping the society's direction and I am excited to see what they accomplish in future, as one of the UK's largest hackathons!
-
-![Photo of the HackSheffield committee](/assets/images/dynamic/hacksheffield/committee.jpg)
-
-I'm the one third from the left, looking ecstatic that HackSheffield 4.0 had such an impressive turn-out and (almost) everything went exactly as planned.
-
-## Hackathons I've Attended
-
-I've attended loads of hackathons since starting university --- you can see some of the projects I've created at hackathons [on Devpost](https://devpost.com/gregives). I decided to create a map to showcase the amazing places I've had the chance to visit for hackathons, including Bordeaux, St Gallen and Barcelona!
-
-<iframe title="Greg's Hackathons – Google My Maps" src="https://www.google.com/maps/d/u/0/embed?mid=1TPa-VDk3oQ1qIfv2efqFSsRUzG5Zj2p0" width="640" height="480"></iframe>
