--- contents/projects/hype.md
@@ -1,25 +1,0 @@
----
-title: Hype
-titleShort: ''
-date: 2016-10-01
-image: '/assets/images/dynamic/hype/screenshots1.png'
-description: Product of my first hackathon. Hype, an Android app to “hype” a location or event, which can be viewed on an interactive map
-tags:
-  - Hackathon
-  - Android
-  - Java
-  - Google Maps
-  - Node.js
-  - Express
-  - MongoDB
----
-
-The first hackathon I attended was HackSheffield 2.0, in my first month at university. I formed a team with three other people from my course, and we started brainstorming something to build at our first hackathon! We decided on an app to share events with other people, with the focus point of the app being an interactive map to view all events happening around you, and "hype" up events which you attended.
-
-![Screenshots of the Hype app](/assets/images/dynamic/hype/screenshots1.png)
-
-I worked on the frontend of the app in Android Studio, designing the user interface and the icon and logo. It was the first time I had created an Android app from scratch and connected it to a backend, which was a great challenge for me. I enjoyed working in such a fast-paced environment --- hackathons are great for learning new stuff! The backend of Hype was written using Express and MongoDB, in order to store the events, their locations and how "hyped" each event was. The more people who hyped an event, the larger the event would show up on the heatmap.
-
-![Hype app icon](/assets/images/dynamic/hype/icon.png '15rem')
-
-We presented our app to more than one hundred attendees at the hackathon and talked about what we had learnt at our first hackathon. I found it an incredible experience and as such a became involved in the organising committee for HackSheffield. You can read more about <nuxt-link to="/projects/hacksheffield/">my role in HackSheffield and the other hackathons I've attended</nuxt-link>.
