--- contents/projects/icon-packs.md
@@ -1,41 +1,0 @@
----
-title: Icon Packs
-titleShort: ''
-date: 2016-08-01
-image: '/assets/images/dynamic/icon-packs/squircle7.png'
-description: My first foray into programming and design. Icon packs for custom Android launchers, so you can make your phone homescreen look pretty
-tags:
-  - Android
-  - Icon packs
-  - Java
-  - Design
-  - Promotion
----
-
-<post-aside type="info">
-
-I have decided to remove my icon packs from the Google Play Store due to the amount of time required for maintenance and responding to users. You can read all about them here though!
-
-</post-aside>
-
-I made my first Android app aged 14 --- I brought my coding and design skills together to create icon packs for custom launchers, such as Nova Launcher. My first icon pack, Trace, was designed as a transparent outlined-icon pack to allow minimal distraction from your homescreen wallpaper.
-
-> Enhance the style of your homescreen with minimal interference from icons --- **Trace** icons are subtle and elegant, whilst allowing your wallpaper to show through.
-
-![Promotional design for Trace icons](/assets/images/dynamic/icon-packs/trace1.png)
-
-Trace contains more than 1,000 icons, all created using Photoshop Touch on a Samsung Note tablet with a stylus!
-
-![Screenshots of Trace on homescreen](/assets/images/dynamic/icon-packs/trace2.png)
-
-After creating Trace, I started creating another icon pack named Squircle. As you might have guessed, the icon pack was centred around its shape --- somewhere between a circle and a square. The squircle shape has now become a very popular base for icons, for example, the default stock icons on Samsung phones.
-
-> Introducing **Squircle**, a simplistic yet beautiful icon pack consisting of the characteristic squircle-shaped coloured outline, flat colours, and a subtle drop shadow. **Bring your home screen to life!**
-
-![Promotional design for Squircle](/assets/images/dynamic/icon-packs/squircle1.png)
-
-![Promotional design for Squircle](/assets/images/dynamic/icon-packs/squircle2.png)
-
-![Promotional design for Squircle](/assets/images/dynamic/icon-packs/squircle3.png)
-
-I decided to offer Squircle as a free app, with a paid "donation" app available for those who wanted to support the development. Squircle has now been downloaded more than 300,000 times and has more than 5,000 ratings.
