--- contents/projects/picnic-spots.md
@@ -1,28 +1,0 @@
----
-title: Picnic Spots
-titleShort: ''
-date: 2019-04-01
-image: '/assets/images/dynamic/picnic-spots/screenshot.png'
-description: Website created at HackBordeaux in 24 hours. View, add and review the best places to eat your tuna sweetcorn sandwiches
-tags:
-  - Node.js
-  - Express
-  - EJS
-  - Passport
-  - Bootstrap
-  - Mapbox
----
-
-Picnic Spots was created at HackBordeaux with another member of HackSheffield, in 24 hours. Picnic Spots allows you to find a perfect spot for your lunch, courtesy of user-submitted spots and reviews.
-
-![Screenshot of Picnic Spots, showing a map with locations and ratings](/assets/images/dynamic/picnic-spots/screenshot.png)
-
-I created Picnic Spots with another team member from HackSheffield, [Arnas Boruchovas](https://arnas.codes). At first we were struggling to come up with an idea, so we took some time outside to have a break. Whilst searching for a place to eat our lunch, our idea strook --- a website to find nice picnic spots!
-
-Picnic Spots was created using:
-
-- Express for routing and middleware
-- EJS for templating
-- Sass for styling
-- Passport for authentication
-- Mapbox for the interactive map
