--- contents/projects/university-of-sheffield-brass-band.md
@@ -1,35 +1,0 @@
----
-title: University of Sheffield Brass Band
-titleShort: USBB
-date: 2019-03-01
-image: '/assets/images/dynamic/usbb/website1.png'
-description: Website for the University of Sheffield Brass Band to showcase concerts, news and photos. Uses Jekyll for easy content management
-tags:
-  - Jekyll
-  - Web design
-  - Bootstrap
-  - Content management
----
-
-I have been a member of the University of Sheffield Brass Band for three years, and decided to create a website for them to promote concerts, news and photos.
-
-![Screenshot of the home page of USBB's website](/assets/images/dynamic/usbb/website2.png)
-
-## Requirements
-
-An essential requirement of the brass band's website was that people with only a little technical experience would be able to edit it, or even people who had no prior experience. The website was designed with four pages in mind:
-
-- The home page aims to provide an overview of the band, with some photos of the band in action, a news section and finally a short description of the band and its function.
-- A concerts page should promote the band's upcoming concerts, providing ticketing information and possibly a banner or graphic designed for the concert.
-- Members of the band should be able to upload images to the gallery page with little configuration.
-- Finally, the contact page allows potential customers to enquire about the band.
-
-## Technology
-
-Jekyll seemed a perfect static site generator for the job as future members of the band would be able to edit the website simply by editing or adding markdown files. News posts and concerts would be easily handled by Jekyll's built in blog functionality.
-
-I designed the website using [Bootstrap](https://getbootstrap.com) and some custom styles, nothing too fancy! The resulting design looks clean and intuitive, whilst maintaining the band's branding. The contact form uses [formcarry](https://formcarry.com), a free backend for static forms, and the footer includes a Facebook banner so people can easily connect with us.
-
-![Screenshot of the concerts page of USBB's website](/assets/images/dynamic/usbb/website3.png)
-
-![Screenshot of the contact page of USBB's website](/assets/images/dynamic/usbb/website4.png)
