+++ contents/services/asset-tracking.md
@@ -1,0 +1,15 @@
+---
+title: Asset Tracking
+titleShort: ''
+date: 2021-02-08
+image: '/assets/images/dynamic/asset-tracking.png'
+description: Tap in to the world of New Space and IoT with AI to know where exactly in the world your assets are now, and where they're likely to be later
+tags:
+  - Eleventy
+  - Dependency
+  - Loaders
+---
+
+![Photo of Asset Tracking](/assets/images/dynamic/asset-tracking.png)
+
+Tap in to the world of New Space and IoT with AI to know where exactly in the world your assets are now, and where they're likely to be later
