+++ contents/services/density-analysis.md
@@ -1,0 +1,18 @@
+---
+title: Density Analysis
+titleShort: ''
+date: 2019-04-01
+image: '/assets/images/dynamic/density-analysis.png'
+description: Get time-series information on the utilisation of airports, shopping malls and ports with our AI aggregating all the data for you
+tags:
+  - Node.js
+  - Express
+  - socket.io
+  - MongoDB
+  - Passport
+  - Vue
+  - Mapbox
+type: hidden
+---
+
+![Photo of Density Analysis](/assets/images/dynamic/density-analysis.png)
