+++ contents/team.md
@@ -1,0 +1,12 @@
+
+<a href=https://uk.linkedin.com/in/sarpulas target=_blank>Sarp Ulas OCAK - CEO </a>
+
+<a href=https://tr.linkedin.com/in/meteseyrek target=_blank>Mete Seyrek - Business Development Manager </a>
+
+<a href=https://www.linkedin.com/in/ezgikayar target=_blank>Ezgi Kayar - UI/UX Designer </a>
+
+<a href=https://www.linkedin.com/in/berkaykarlik target=_blank>Berkay Karlik - Software Engineer </a>
+
+<a href=https://www.linkedin.com/in/furkanaldemir target=_blank>Furkan Aldemir - Software Engineer </a>
+
+<a href=https://www.linkedin.com/in/utku-kalkanli target=_blank>Utku Kalkanli - Software Engineer </a>
