--- contents/watches.md
@@ -1,1 +1,0 @@
-**Welcome to my new blog about wristwatches!** Over the last few years I've started collecting watches on a budget; read my review of the watches I've bought and find out how well they fit my small 6&#8209;inch&nbsp;wrists.
