+++ pages/blog/_post.vue
@@ -50,12 +50,12 @@ export default {
         {
           hid: 'og:url',
           property: 'og:url',
-          content: `https://gregives.co.uk${this.post.link}`
+          content: `https://novit.ai${this.post.link}`
         },
         {
           hid: 'og:image',
           property: 'og:image',
-          content: `https://gregives.co.uk${this.post.link}/og.png`
+          content: `https://novit.ai${this.post.link}/og.png`
         }
       ]
     }
